[FOLD] Simplify Validations::add:
In the case when a validator switched signing keys, the old code tried
to detect a case where the new validator re-issued the same sequence
validation with the new key, but for a different ledger. This is not
ideal, since there is no gaurantee when other validators see the two
different validations or the manifest updates.

The next changeset requiring increasing full validation sequence numbers
will makes this change more relevant, so updating tests will be deferred
until then.